aerospace / qutebrowser: move cursor-related keymaps to aerospace
All of a sudden calling cursor.swift from within qutebrowser ceases to
work. To be more precise, the script runs successfully but the cursor
refuses to be manipulated. The bug is fixed by moving the cursor
manipulation keymaps to aerospace. Moreoever, this makes more sense as
it is now possible to globally move the cursor via the keyboard.